Prevent sqoop from escaping the null string.
This is useful for working with hive. The null character in hive is \N.
Without this option sqoop converts to \\N. By default I think that that
this should be enabled, but I kept it as an option to preserve backwards
compatibility. Here is the option to pass into sqoop to enable this
funcationality:

--do-not-escape-null-string